fix:crud user et group

+++ src/controllers/group_controller.js
@@ -10,7 +10,7 @@ exports.list = async (req, res) => {
         const list = await Group.find({}, null, {
             skip: firstIndex,
             limit: perPage
-        }).populate("admin");
+        }).populate("admin").populate("members");
 
         res.json(list);
     }
@@ -30,11 +30,11 @@ exports.read = async (req, res) => {
             populate: {
                 path: "groups"
             }
-        });
+        }).populate("members");
 
         if (!group){
             res.status(404).json({
-                message: "NOT FOUND"
+                message: "group not found"
             });
             return;
         }
@@ -59,7 +59,7 @@ exports.insert = async (req, res) => {
         user.groups.push(saved);
         await user.save();
         res.json({
-            message: "OK",
+            message: `group ${group.name} inserted`,
             data: {
                 "_id": saved._id,
                 "name": saved.name
@@ -77,9 +77,7 @@ exports.delete = async (req, res) => {
     const id = req.params.id;
 
     try {
-        const group = await Group.findById(id).populate({
-            path: "admin"
-        });
+        const group = await Group.findById(id).populate("admin").populate("members");
 
         if (!group){
             res.status(404).json({
@@ -87,12 +85,15 @@ exports.delete = async (req, res) => {
             });
             return;
         }
-        const user = await User.findById(group.admin.id).populate("groups");
+        if (group.members.length){
+            throw new Error(`the group has members`);
+        }
+        const user = await User.findById(group.admin._id).populate("groups");
         user.groups = user.groups.filter(group => group._id != id);
         await user.save();
-        const status = await group.delete()
+        const status = await group.delete();
         res.json({
-            message: "OK"
+            message: `group ${group.name} deleted`
         });
     }
     catch (e){
@@ -106,13 +107,66 @@ exports.update = async (req, res) => {
     const body = req.body;
     const id = req.params.id;
     body._id = id;
+    if (!body.members){
+        body.members = [];
+    }
+    const data_members = body.members.slice();
+    body.members = [];
     const group = new Group(body);
 
     try {
         const updated = await Group.findByIdAndUpdate(id, group);
+        updated.members = data_members;
+        await updated.save();
+        const members = updated.members;
+        const users = await User.find({
+            groups: [id]
+        });
 
+        for (let i = 0; i < users.length; i++){
+            const user = users[i];
+            let flag = true;
+
+            for (let member of members){
+                if (user._id.toString() == member.toString()){
+                    flag = false;
+                }
+            }
+            if (user._id.toString() == updated.admin.toString()){
+                console.log("enter");
+            }
+            else if (flag){
+                let f2 = true;
+
+                while (f2){
+                    const index = user.groups.findIndex(el => el.toString() == id.toString());
+
+                    if (index > -1){
+                        user.groups.splice(index, 1);
+                    }
+                    else {
+                        f2 = false;
+                    }
+                }
+                await user.save();
+            }
+        }
+        for (let member of members){
+            const user = await User.findById(member);
+            let flag = true;
+
+            for (let group of user.groups){
+                if (group.toString() == id.toString()){
+                    flag = false;
+                }
+            }
+            if (flag){
+                user.groups.push(id);
+                await user.save();
+            }
+        }
         res.json({
-            message: "OK"
+            message: `group ${updated.name} updated`
         });
     }
     catch (e){
